requirements.txt
aiohttp==3.2.0
aiomysql==0.0.14
aiosocksy==0.1
aiotg==0.9.9
argh==0.26.2
async-timeout==3.0.0
asyncio==3.4.3
attrs==18.1.0
cairocffi==0.8.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cssselect2==0.2.1
defusedxml==0.5.0
Django==2.0.5
idna==2.6
lxml==4.2.1
multidict==4.3.1
mysqlclient==1.3.12
numpy==1.14.3
pathtools==0.1.2
Pillow==5.1.0
pycparser==2.18
PyMySQL==0.8.1
PySocks==1.6.8
pytz==2018.4
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
SQLAlchemy==1.2.7
svglib==0.8.1
tinycss2==0.6.1
urllib3==1.22
watchdog==0.8.3
webencodings==0.5.1
yarl==1.2.3
